+++ .editorconfig
@@ -6,26 +6,23 @@ trim_trailing_whitespace = true
 end_of_line = lf
 
 [*.{kt,kts,md}]
+ij_kotlin_code_style_defaults = KOTLIN_OFFICIAL
+ij_kotlin_allow_trailing_comma = true
+ij_kotlin_allow_trailing_comma_on_call_site = true
+ij_kotlin_imports_layout = *,java.**,javax.**,kotlin.**,^
+ij_kotlin_packages_to_use_import_on_demand = kotlinx.android.synthetic.**
+ij_kotlin_name_count_to_use_star_import = 100
+ij_kotlin_name_count_to_use_star_import_for_members = 100
 indent_size = 4
-continuation_indent_size = 4
 indent_style = space
 insert_final_newline = true
-
-# ktlint styles
-ij_kotlin_code_style_defaults = KOTLIN_OFFICIAL
-ij_kotlin_line_comment_at_first_column = false
-ij_kotlin_line_comment_add_space = true
-ij_kotlin_name_count_to_use_star_import = 2147483647
-ij_kotlin_name_count_to_use_star_import_for_members = 2147483647
-ij_kotlin_keep_blank_lines_in_declarations = 1
-ij_kotlin_keep_blank_lines_in_code = 1
-ij_kotlin_keep_blank_lines_before_right_brace = 0
-ij_kotlin_align_multiline_parameters = false
-ij_continuation_indent_size = 4
-ij_kotlin_import_nested_classes = false
-ij_kotlin_imports_layout = *, ^
-ktlint_standard_trailing-comma-on-declaration-site=disabled
-ktlint_standard_trailing-comma-on-call-site=disabled
+ktlint_code_style = official
+ktlint_function_signature_body_expression_wrapping = default
+ktlint_function_signature_rule_force_multiline_when_parameter_count_greater_or_equal_than = -1
+ktlint_ignore_back_ticked_identifier = false
+ktlint_no-wildcard-imports = disabled
+ktlint_standard_no-wildcard-imports = disabled
+max_line_length = 120
 
 [**/test/**.kt]
 ktlint_ignore_back_ticked_identifier = true
