+++ examples/kotlin/src/main/kotlin/ExactSizeQRCode.kt
@@ -1,10 +1,10 @@
+import qrcode.raw.QRCodeBuilder
 import java.io.FileOutputStream
-import qrcode.QRCode
 
 class ExactSizeQRCode {
     fun createQRCode(content: String, size: Int) {
         FileOutputStream("kotlin-exact-$size.png").use {
-            val qrCode = QRCode(content)
+            val qrCode = QRCodeBuilder(content)
             val qrCodeData = qrCode.encode()
 
             // qrCodeData[0].size = How many columns (X axis) the finished QRCode will have
@@ -18,7 +18,7 @@ class ExactSizeQRCode {
                 // Many times the image will be a bit smaller or a bit bigger because sizePerCell is an Int
                 // This adds a few TRANSPARENT pixels of border around it (if smaller)
                 // Or reduces the image by however many pixels needed, so it'll be the exact size
-                val adjustedImage = qrCode.qrCodeGraphicsFactory.newGraphics(size, size)
+                val adjustedImage = qrCode.graphicsFactory.newGraphics(size, size)
                 val adjustedX = (size / 2) - (renderedImage.width / 2)
                 val adjustedY = (size / 2) - (renderedImage.width / 2)
