+++ examples/kotlin/src/main/kotlin/QRCodeKotlinLogo.kt
@@ -1,3 +1,12 @@
+import qrcode.color.Colors
+import qrcode.internals.QRCodeRegion.BOTTOM_LEFT_CORNER
+import qrcode.internals.QRCodeRegion.BOTTOM_RIGHT_CORNER
+import qrcode.internals.QRCodeRegion.TOP_LEFT_CORNER
+import qrcode.internals.QRCodeRegion.TOP_RIGHT_CORNER
+import qrcode.internals.QRCodeSquareType.MARGIN
+import qrcode.internals.QRCodeSquareType.POSITION_PROBE
+import qrcode.raw.QRCodeBuilder
+import qrcode.render.QRCodeGraphics
 import java.awt.Color
 import java.awt.MultipleGradientPaint.CycleMethod
 import java.awt.Polygon
@@ -8,15 +17,6 @@ import java.awt.image.BufferedImage
 import java.io.FileOutputStream
 import javax.imageio.ImageIO
 import kotlin.math.roundToInt
-import qrcode.Colors
-import qrcode.QRCode
-import qrcode.internals.QRCodeRegion.BOTTOM_LEFT_CORNER
-import qrcode.internals.QRCodeRegion.BOTTOM_RIGHT_CORNER
-import qrcode.internals.QRCodeRegion.TOP_LEFT_CORNER
-import qrcode.internals.QRCodeRegion.TOP_RIGHT_CORNER
-import qrcode.internals.QRCodeSquareType.MARGIN
-import qrcode.internals.QRCodeSquareType.POSITION_PROBE
-import qrcode.render.QRCodeGraphics
 
 /**
  * This is a special, very custom example. It only works on the JVM. It uses the AWT classes directly to create a very
@@ -29,7 +29,7 @@ class QRCodeKotlinLogo {
 
     fun createQRCodeKotlinLogo(cellSize: Int, margin: Int = (cellSize / 1.16).roundToInt()) {
         val renderedQrCodeImage =
-            QRCode("QRCode Kotlin").renderShaded(cellSize, margin) { qrCodeSquare, qrCodeGraphics ->
+            QRCodeBuilder("QRCode Kotlin").renderShaded(cellSize, margin) { qrCodeSquare, qrCodeGraphics ->
                 if (qrCodeSquare.dark) {
                     when (qrCodeSquare.squareInfo.type) {
                         POSITION_PROBE -> when (qrCodeSquare.squareInfo.region) {
