+++ examples/kotlin/src/main/kotlin/SimpleQRCode.kt
@@ -1,10 +1,10 @@
+import qrcode.raw.QRCodeBuilder
 import java.io.FileOutputStream
-import qrcode.QRCode
 
 class SimpleQRCode {
     fun createQRCode(content: String) {
         FileOutputStream("kotlin-simple.png").use {
-            QRCode(content).render().writeImage(it)
+            QRCodeBuilder(content).render().writeImage(it)
         }
     }
 }
