+++ examples/kotlin/src/main/kotlin/ZXingQRCode.kt
@@ -1,7 +1,7 @@
+import qrcode.raw.ErrorCorrectionLevel
+import qrcode.raw.MaskPattern
+import qrcode.raw.QRCodeBuilder
 import java.io.FileOutputStream
-import qrcode.ErrorCorrectionLevel
-import qrcode.MaskPattern
-import qrcode.QRCode
 
 /**
  * Tries to generate a QRCode as close as possible to the one generated by the following URL:
@@ -22,7 +22,7 @@ class ZXingQRCode {
     fun createQRCode(content: String) {
         val cellSize = 12
         val margin = 49
-        val qrCode = QRCode(content, ErrorCorrectionLevel.M)
+        val qrCode = QRCodeBuilder(content, ErrorCorrectionLevel.M)
         val qrCodeData = qrCode.encode(type = 1, maskPattern = MaskPattern.PATTERN000)
 
         val fileOut = FileOutputStream("kotlin-zxing-hello-world.png")
