+++ examples/kotlin/src/main/kotlin/fancy/HelloCirclesColors.kt
@@ -1,11 +1,11 @@
 package fancy
 
+import qrcode.QRCode
+import qrcode.color.Colors
 import java.io.FileOutputStream
-import qrcode.Colors
-import qrcode.fancy.FancyQRCode
 
 fun main() {
-    val qrCode = FancyQRCode.newCircleBuilder()
+    val qrCode = QRCode.ofCircles()
         .withColor(Colors.DODGER_BLUE)
         .build("Hello, circles... with color!")
