+++ src/commonMain/kotlin/qrcode/internals/QRMath.kt
@@ -28,6 +28,18 @@ internal object QRMath {
         return EXP_TABLE[i]
     }
 
+    /**
+     * From https://math.stackexchange.com/a/3086626/874981
+     */
+    fun rectsIntersect(x1: Int, y1: Int, w1: Int, h1: Int, x2: Int, y2: Int, w2: Int, h2: Int): Boolean {
+        val x1End = x1 + w1
+        val y1End = y1 + h1
+        val x2End = x2 + w2
+        val y2End = y2 + h2
+
+        return x1End < x2 || x1 > x2End || y1End < y2 || y1 > y2End
+    }
+
     init {
         for (i in 0..7) {
             EXP_TABLE[i] = 1 shl i
